Expose lookup
Hi, I recently came across another usage of the router which is to have it parse routes and give you back the params it matched instead of having it run the callback itself. `lookup` needs to be exposed and that's what this PR is all about :). What do you think about it? Thanks, Darío