Separate query params from route params
This is a breaking change. Four arguments are returned to the route
handler instead of three. All query parameters are passed in the new
second argument, instead of being merged with the route parameters in
the first argument. The other two arguments are unchanged.

Tests and documentation are updated and version is set to 3.0.0, as this
changes the API.

Minified size is reduced to 689.